% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EHUM.R
\name{estimate_EHUM}
\alias{estimate_EHUM}
\title{Empirical Hyper Volume Under Manifolds}
\usage{
estimate_EHUM(beta, labels, x_mat)
}
\arguments{
\item{beta}{The parameter we measure EHUM based on.}

\item{labels}{The labels of the Columns of the data matrix.}

\item{x_mat}{The Data Matrix}
}
\value{
Empirical Hyper-volume Under Maniforlds Estimate
}
\description{
An estimator of Hyper Volume Under Manifolds
}
\examples{
estimate_EHUM(rep(1,12),colnames(AL),AL)


estimate_EHUM(1:10 , sample(c(rep("lab1",10),rep("lab2",10),rep("lab3",10))), matrix(rnorm(300), nrow = 10))

}
