% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ULBA.R
\name{estimate_ULBA}
\alias{estimate_ULBA}
\title{Upper And Lower Bound Approach}
\usage{
estimate_ULBA(beta, labels, x_mat)
}
\arguments{
\item{beta}{The parameter we measure ULBA based on.}

\item{labels}{The labels of the Columns of the data matrix.}

\item{x_mat}{The Data Matrix}
}
\value{
Upper and Lower Bound Approach on empirical Hyper-volume Under Manifolds Estimate
}
\description{
`ULBA` is an another approach to Hyper Volume Under Manifold Problem
}
\examples{
estimate_ULBA(rep(1,12),colnames(AL),AL)


estimate_ULBA(1:10 , sample(c(rep("lab1",10),rep("lab2",10),rep("lab3",10))), matrix(rnorm(300), nrow = 10))


}
