% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCOR-package.R
\docType{package}
\name{SCOR-package}
\alias{SCOR}
\alias{SCOR-package}
\title{SCOR: Spherically Constrained Optimization Routine}
\description{
A non convex optimization package that optimizes any function under the criterion, combination of
              variables are on the surface of a unit sphere, as described in the paper : Das et al. (2019) <arXiv:1909.04024> .
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/synx21/SCOR}
}

}
\author{
\strong{Maintainer}: Debsurya De \email{debsurya001@gmail.com}

Authors:
\itemize{
  \item Priyam Das \email{pdas@ncsu.edu}
}

}
\keyword{internal}
