% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scor.R
\name{scor}
\alias{scor}
\title{Main function for ScoR}
\usage{
scor(otu_tab, sparCor, sparP, zc = 0.2)
}
\arguments{
\item{otu_tab}{(Required). File path for the OTU table. Remove \verb{#} from the first cell.}

\item{sparCor}{(Required). File path for SparCC correlation matrix. Remove \verb{#} from the first cell.}

\item{sparP}{(Required). File path for SparCC p-value matrix. Remove \verb{#} from the first cell.}

\item{zc.}{Fraction of zero-counts allowed for each pair of OTUs. Default = 0.1.
OTUs which have more zeros than zc will be excluded.}
}
\value{
A \code{list} object.
}
\description{
Creates a \code{list} output which includes an igraph object,
a module count, r threshold value for SparCC correlation,
r threshold for distance correlation and
a final correlation table in long format.
}
\examples{
data(enterotype)
sample_data(enterotype)
EThealthy <- FSprep(enterotype, ClinicalStatus==healthy, prev = 0.2)
write.table(EThealthy, sep ="\t", file = "Healthy.txt", quote = FALSE,row.names = FALSE)
#RUN FastSpar to generate SparCC correlation and p-value.
sr <- scor("Healthy.txt","H_cor.txt","H_p.txt",0.2)

}
\author{
Abhishake Lahiri \email{abhishake.lahiri@gmail.com}
}
