% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youden_points.R
\name{youden_points}
\alias{youden_points}
\title{Finding Youden Indices}
\usage{
youden_points(beta, labels, x_mat, grid_size = 100)
}
\arguments{
\item{beta}{The parameter we do HUM based on}

\item{labels}{The labels of the Columns of the data matrix.}

\item{x_mat}{The Data Matrix}

\item{grid_size}{The size of increment in the grid we check cutpoints against. Default value is 100.}
}
\value{
Youden Indices and Cut Points
}
\description{
A function to find Youden Indices and Cutpoints for number of categories less than equal to 3.
}
\examples{

beta <- c(-0.399,-0.155,-0.265,-0.184,
     -0.267,0.666,-0.187,0.273,0.0463,0.167,0.163,0.178)

youden_points(beta,colnames(AL),AL)


}
