% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YoudenBoxPlot.R
\name{YoupointsBoxPlot}
\alias{YoupointsBoxPlot}
\title{Visualization Based On Youden Indices.}
\usage{
YoupointsBoxPlot(beta, labels, x_mat, cat_names = NULL,
  grid_size = 100)
}
\arguments{
\item{beta}{The parameter we do HUM based on}

\item{labels}{The labels of the Columns of the data matrix}

\item{x_mat}{The Data Matrix}

\item{cat_names}{The vector of strings containing category names.}

\item{grid_size}{The size of increment in the grid we check cutpoints against. Default value is 100.}
}
\value{
Box Plot Visualization Based On Youden Indices
}
\description{
A Box Plot Visualization Based On Youden Indices for less than equal to 3 categories.
}
\examples{

beta <- c(-0.399,-0.155,-0.265,-0.184,
     -0.267,0.666,-0.187,0.273,0.0463,0.167,0.163,0.178)

YoupointsBoxPlot(beta,colnames(AL),AL, cat_names = c("Healthy","MCI","AD"))
}
