% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCOR-package.R
\docType{package}
\name{SCOR-package}
\alias{SCOR}
\alias{SCOR-package}
\title{SCOR: Spherically Constrained Optimization Routine}
\description{
A package to solve optimization problem where parameters belong to the surface of unit hyper-sphere.This package also includes functions to obtain optimal coefficients for combining multiple bio-markers for classification of multi-class disease outcome categories, using different criteria namely Empirical Hyper-volume Under Maniforlds (EHUM), Smooth approximation of the empirical Hyper-volume Under Maniforlds (SHUM) and Upper and Lower Bound Approach on empirical hyper-volume under maniforlds (ULBA).
}
\author{
\strong{Maintainer}: Debsurya De \email{debsurya001@gmail.com}

Authors:
\itemize{
  \item Priyam Das \email{pdas@ncsu.edu}
}

}
\keyword{internal}
